Split command palette focus restore coordinator
